requirements.txt
apache-airflow>=2.0.0
flask>=1.1.0
flask-admin>=1.5.0
sqlalchemy>=1.3.0
requests>=2.25.0
jinja2>=2.11.0
